Add files via upload
Online product recommendation is a sales strategy for shopping platforms. It is a way to improve both sales and the client shopping experience. An online client gets recommendations based on what other customers that have the same shopping tendencies as him bought. This means that to be recommended, a product needs to be purchased at least once. Our methods are based on the buyer`s shopping history. The idea is to recommend products that resembles the ones he bought. Techniques such as the “K nearest neighbors”or convolutional neural networks are used to do so.  For example, if a ring is chosen,the algorithm gives images of five other rings of the same style or recommends a matching neckless or earrings to a customer that purchased a ring.